+++ CHANGELOG.md
@@ -9,7 +9,8 @@
 - Added language formatter [#1410](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1410)
 - Added default file template [#1411](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1411)
 - Added language folding support [#1412](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1412)
-- Added reference resolution for Type column [#1413](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1413)
+- Added reference resolution for `Type` column [#1413](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1413)
+- Added reference resolution for `Target` column [#1414](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1414)
 
 ## [2025.1.4]